+++ docs/coreComponents/ModalForm/innerRef/index.tsx
@@ -6,8 +6,8 @@ import { useRef, useState } from 'react';
 const InnerRefDemo = () => {
   const ref = useRef();
 
-  const onFinish = (values: any) => {
-    console.log({ values });
+  const onFinish = (values: any, formType, formData) => {
+    console.log({ formType, values, formData });
   };
 
   return (
@@ -15,7 +15,7 @@ const InnerRefDemo = () => {
       <div style={{ textAlign: 'start' }}>
         <Button
           type="primary"
-          onClick={() => ref.current?.openModal({ username: '王先生', phone: 134 })}
+          onClick={() => ref.current?.openModal('new', { username: '王先生', phone: 134 })}
         >
           打开弹窗
         </Button>