+++ docs/coreComponents/ProTable/readonly/apis.ts
@@ -1,0 +1,14 @@
+export const mockRequest = () => {
+  return new Promise((resolve) => {
+    setTimeout(
+      () =>
+        resolve({
+          data: [
+            { id: 1, name: 'Jack', nickName: 'Jackiy', phone: 133, sex: '男' },
+            { id: 2, name: 'Tom', phone: 155 },
+          ],
+        }),
+      1000,
+    );
+  });
+};