+++ CHANGELOG.md
@@ -73,3 +73,6 @@
 
 - ModalForm onFinish 失败后不关闭弹窗
 - 修复 BusinessSelectBuilder 分页组件没有抛出 queryParams
+- ModalForm 的 innerRef 的 openModal 方法增加 formType 参数;onOpen 和 onFinish 时可获取到; formType 支持 read
+- ProTable 的 innerRef 的 openModal 支持 read
+- 修复 ProTable 的 columnsType